src/data_loader.py
import json

class DataLoader:
    """Loads and preprocesses data from a JSON file.

    Attributes:
        file_path (str): The path to the JSON file.
    """

    def __init__(self, file_path):
        """Initializes the DataLoader with the given file path.

        Args:
            file_path (str): The path to the JSON file.
        """
        self.file_path = file_path

    def load_data(self):
        """Loads data from the JSON file.

        Returns:
            list: A list of dictionaries, where each dictionary represents a data sample.
        """
        try:
            with open(self.file_path, 'r') as f:
                data = json.load(f)
            return data
        except FileNotFoundError:
            print(f"Error: File not found at {self.file_path}")
            return []
        except json.JSONDecodeError:
            print(f"Error: Invalid JSON format in {self.file_path}")
            return []

    def preprocess_data(self, data):
        """Preprocesses the loaded data.

        Args:
            data (list): A list of dictionaries representing the data.

        Returns:
            list: A list of preprocessed dictionaries.
        """
        # Implement preprocessing logic here (e.g., tokenization, cleaning)
        # This is a placeholder; replace with actual preprocessing steps
        return data